worker/requirements.txt
redis==4.6.0        # ← change from 5.0.7 to 4.6.0
pymongo==4.8.0
python-dotenv==1.0.1